Broke out API usage into module and added AWS SageMaker support
